requirements.txt
bs4>=0.0.2
datasets>=3.6.0
langchain>=0.3.26
langchain-community>=0.3.26
langchain-openai>=0.3.25
ragas>=0.2.15
rapidfuzz>=3.13.0
scikit-learn>=1.7.0
pandas
python-dotenv